+++ .gitignore
@@ -1,6 +1,4 @@
 .idea/
-composer.json
 composer.phar
 package.json
-composer.lock
 vendor/
