+++ app/public/css/style.css
@@ -3,5 +3,5 @@ main {
 }
 
 footer {
-    min-height: 130px;
+    min-height: 160px;
 }
