+++ app/src/controller/Controller.php
@@ -8,28 +8,27 @@
 
 namespace app\src\controller;
 
+use app\src\dao\CategoryDAO;
 use app\src\dao\CommentDAO;
 use app\src\dao\PostDAO;
 use app\src\dao\UserDAO;
-use app\src\dao\CategoryDAO;
-use app\src\model\View;
 use app\src\model\ViewTwig;
 
 class Controller
 {
-    protected $view;
-    protected $postDAO;
-    protected $commentDAO;
-    protected $userDAO;
     protected $categoryDAO;
+    protected $commentDAO;
+    protected $postDAO;
+    protected $userDAO;
+    protected $view;
     protected $viewTwig;
 
     public function __construct()
     {
-        $this->postDAO = new PostDAO();
-        $this->commentDAO = new CommentDAO();
-        $this->userDAO = new UserDAO();
         $this->categoryDAO = new CategoryDAO();
+        $this->commentDAO = new CommentDAO();
+        $this->postDAO = new PostDAO();
+        $this->userDAO = new UserDAO();
         $this->viewTwig = new ViewTwig();
     }
 }
