+++ app/src/controller/UserController.php
@@ -1,111 +1,22 @@
 <?php
+/**
+ * Created by PhpStorm.
+ * User: Alex
+ * Date: 27/11/2018
+ * Time: 13:14
+ */
 
 namespace app\src\controller;
 
 use app\core\Utils;
 
-class BackController extends Controller
+class UserController extends Controller
 {
-    public function getAdmin()
-    {
-        if (Utils::isAdmin()) {
-            $posts = $this->postDAO->getAll('all', null, true);
-            $comments = $this->commentDAO->getComments('waiting');
-            $users = $this->userDAO->getAllUser();
-            $this->viewTwig->render('admin.twig', ['posts' => $posts, 'comments' => $comments, 'users' => $users]);
-        } else {
-            header('Location: index.php');
-        }
-    }
-
-    public function addPost($post)
-    {
-        if (Utils::isAdmin()) {
-            if (isset($post['submit'])) {
-                if (Utils::checkField(['title', 'chapo', 'content', 'category'], $post)) {
-                    $post = $this->postDAO->add($post, $_SESSION['userId']);
-                    $successM = 'Votre nouvel article à bien été créé !';
-                    $message = Utils::messageAlert(true, $successM, null);
-                    Utils::addFlashBag('message', $message);
-                    header('Location: index.php?p=post&idArt=' . $post);
-                } else {
-                    $errorM = 'Veuillez définir tout les champs du formulaire !';
-                    $message = Utils::messageAlert(false, null, $errorM);
-                    Utils::addFlashBag('message', $message);
-                }
-            }
-            $category = $this->categoryDAO->getCategory();
-            $this->viewTwig->render('add_post.twig', ['category' => $category]);
-        }
-    }
-
-    public function updatePost($id, $post = null, $publish = null)
-    {
-        if (Utils::isAdmin()) {
-            if ((isset($post['submit'])) || isset($publish)) {
-                // on vérifie que tout les champs du formulaire sont présent et non vide.
-                if (Utils::checkField(['title', 'chapo', 'content', 'category'], $post) || isset($publish)) {
-                    // si la variable publish est présente
-                    if (isset($publish)) {
-                        //  on la modifie en fonction du parametre passé
-                        $this->postDAO->update($id, null, $publish);
-                        header('Location: index.php?p=admin');
-                    } else {
-                        // si publish n'est pas présent on modifie le contenu de l'article
-                        $this->postDAO->update($id, $post);
-                        header('Location: index.php?p=post&idArt=' . $id);
-                    }
-                } else {
-                    $errorM = 'Veuillez remplir tout les champs du formulaire !';
-                    $successM = null;
-                    $message = Utils::messageAlert(false, $successM, $errorM);
-                    Utils::addFlashBag('message', $message);
-                }
-            }
-            // si la session admin est connecté mais qu'il n'y a pas de soumission de formulaire
-            // on appel le l'article pour remplir le formulaire de modification
-            $post = $this->postDAO->getPost($id);
-            $category = $this->categoryDAO->getCategory();
-            $this->viewTwig->render('update_post.twig', ['post' => $post,
-                'category' => $category
-            ]);
-        }
-    }
-
-    public function deletePost($id)
-    {
-        if (Utils::isAdmin()) {
-            if (isset($id)) {
-                $id = (int)$id;
-                $post = $this->postDAO->delete($id);
-                $errorM = 'L\'article n\'a pu être supprimé !';
-                $successM = 'L\'article à été supprimé !';
-                $message = Utils::messageAlert($post, $successM, $errorM);
-                Utils::addFlashBag('message', $message);
-                header('Location: index.php?p=admin');
-            }
-        }
-    }
-
-    public function getCommentList()
-    {
-        if (Utils::isAdmin()) {
-            $comments = $this->commentDAO->getComments('waiting');
-            $this->viewTwig->render('manage_comment.twig', ['comments' => $comments]);
-        }
-    }
-
-    public function publishComment($idComment, $publish)
-    {
-        if (Utils::isAdmin()) {
-            $this->commentDAO->update($idComment, null, $publish);
-            $successM = 'Le commentaire vient d\'être publié';
-            $message = Utils::messageAlert(true, $successM, null);
-            Utils::addFlashBag('message', $message);
-            header('Location: index.php?p=manage_comment');
-        }
-    }
-
+    /**
+     * @param $post
+     * @param null $userId
+     * @param null $key
+     */
     public function registration($post, $userId = null, $key = null)
     {
         if (isset($post['submit'])) {
@@ -173,30 +84,11 @@ class BackController extends Controller
         }
     }
 
-    public function accountUpdate($id, $quality)
-    {
-        if (Utils::isAdmin()) {
-            if (isset($id) && isset($quality)) {
-                $user = $this->userDAO->getUser('id', $id);
-                $this->userDAO->update($id, $user->getPassword(), $user->getEmail(), $quality,
-                    $user->getStatus(), $user->getValidationKey());
-                Utils::messageSuccess('Compte mis à jour avec succès', 'admin');
-            }
-        }
-    }
-
-    public function addCategory($post)
-    {
-        if (Utils::checkField(['name'], $post)) {
-            $category = $this->categoryDAO->add($post);
-            if ($category) {
-                Utils::messageSuccess('Catégorie ajouté avec succès', 'admin');
-            } else {
-                Utils::messageError('Une erreur c\'est produite veuillez recommencer.', 'admin');
-            }
-        }
-    }
-
+    /**
+     * @param $post
+     * @param null $userId
+     * @param null $key
+     */
     public function passwordUpdate($post, $userId = null, $key = null)
     {
         if (isset($post['submit']) && isset($post['password']) && isset($post['password2'])) {
@@ -227,6 +119,9 @@ class BackController extends Controller
         }
     }
 
+    /**
+     * @param $post
+     */
     public function passwordReset($post)
     {
         if (isset($post['submit']) && isset($post['email'])) {
@@ -248,6 +143,10 @@ class BackController extends Controller
         }
     }
 
+    /**
+     * @param $userId
+     * @param $status
+     */
     public function updateUserStatus($userId, $status)
     {
         if (Utils::isAdmin()) {
@@ -266,6 +165,10 @@ class BackController extends Controller
         }
     }
 
+    /**
+     * @param $userId
+     * @param $quality
+     */
     public function updateUserQuality($userId, $quality)
     {
         if (Utils::isAdmin()) {
@@ -283,22 +186,35 @@ class BackController extends Controller
         }
     }
 
-    public function addNewsletter($post)
+    public function login($login)
     {
-        if (isset($post['submit'])) {
-            if (Utils::checkField(['email'], $post)) {
-                if ((filter_var($post['email'], FILTER_VALIDATE_EMAIL))
-                    && (!$this->userDAO->getNewsletter($post['email']))) {
-                    $this->userDAO->addNewsletter($post['email']);
-                    header('Location: index.php');
+
+        // s'il n'y a pas de session avec un identifiant. on redirige vers le formulaire de connexion.
+        if (!isset($_SESSION['userId'])) {
+            /*si le champ submit est présent et que les champs username et password correspondent,
+        on connecte l'utilisateur*/
+            if (isset($login['submit'])) {
+                if ($this->userDAO->getLogin($login['username'], $login['password'])) {
+                    Utils::messageSuccess('Bienvenue ' . ucfirst($_SESSION['username']) . ' !', 'home');
                 } else {
-                    header('Location: index.php');
+                    $errorM = 'Identifants incorrect ou compte inactif.';
+                    $message = Utils::messageAlert(false, null, $errorM);
+                    Utils::addFlashBag('message', $message);
                 }
-            } else {
-                header('Location: index.php');
             }
+            $this->viewTwig->render('login.twig', []);
         } else {
+            $errorM = ucfirst($_SESSION['username']) . ' vous êtes déjà connecté !';
+            $message = Utils::messageAlert(false, null, $errorM);
+            Utils::addFlashBag('message', $message);
             header('Location: index.php');
         }
     }
+
+    public function logout()
+    {
+        session_destroy();
+        Utils::messageSuccess('Vous êtes Déconnecté !', 'home');
+    }
+
 }
