+++ app/templates/extension/TwigExtension.php
@@ -5,11 +5,13 @@
  * Date: 23/11/2018
  * Time: 23:03
  */
+
 namespace app\templates\extension;
+
 use app\src\dao\CategoryDAO;
 use app\src\dao\PostDAO;
-use Twig_SimpleFunction;
 use app\core\Utils;
+use Twig_SimpleFunction;
 
 class TwigExtension extends \Twig_Extension
 {
@@ -25,13 +27,15 @@ class TwigExtension extends \Twig_Extension
     public function getCategory()
     {
         $categoryDAO = new CategoryDAO();
-        return $category = $categoryDAO->getCategory();
+
+        return $categoryDAO->getCategory();
     }
 
     public function getLastPost()
     {
         $postDAO = new PostDAO();
-        return $post = $postDAO->getLastPost( 4);
+
+        return $postDAO->getLastPost( 4);
     }
 
     public function echoFlashBag()
