+++ app/templates/manage_comment.twig
@@ -5,7 +5,7 @@
 {% block content %}
     <div class="text-center">
         <h2>Gestionnaire des commentaires </h2>
-    </div><br>
+    </div>
     <table class="table table-bordered">
     <tr>
         <th class="text-center" scope="col">Id</th>
@@ -38,5 +38,7 @@
                 </div>
             </th>
         </tr>
+
     {% endfor %}
+    </table>
 {% endblock %}
